requirements.txt
# Core Dependencies
python-binance==1.0.19
pandas==2.2.6
numpy==2.2.6
torch==2.1.2
scikit-learn==1.6.1

# Market Data & Analysis
ccxt==4.2.25
ta-lib==0.4.28
plotly==5.17.0
matplotlib==3.10.3
seaborn==0.13.2

# Data Management
redis==5.0.1
sqlite3
pymongo==4.6.1

# Async & WebSocket
asyncio
websockets==15.0.1
aiohttp==3.12.9

# CoinAPI.io Dependencies
aiofiles==24.1.0
requests==2.32.3

# Configuration & Logging
pyyaml==6.0.2
python-dotenv==1.1.0
structlog==25.4.0

# Testing & Development
pytest==8.4.0
pytest-asyncio==1.0.0
jupyter==1.1.1
ipykernel==6.29.5

# Performance & Monitoring
psutil==7.0.0
memory-profiler==0.61.0

# Rich CLI interface
rich==14.0.0

# Web Dashboard
flask==3.1.1
flask-socketio==5.5.1

# Risk Management
scipy==1.15.3

# Real-time Processing
numba==0.60.0 